requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.24.0
python-docx>=1.0.0
